Change pin variable types from int to int8_t

+++ usermods/usermod_v2_TTGO_TDISPLAY_OUTPUT/TTGO_TDISPLAY_OUTPUT.cpp
@@ -40,17 +40,16 @@ class TTGO_TDISPLAY_OUTPUT : public Usermod {
     uint16_t displayHeight = 170;
     int16_t colOffset = 35;
     
-    int pinMosi = -1;
-    int pinSclk = -1;
-    int pinCs   = -1;
-    int pinDc   = -1;
-    int pinRst  = -1;
-    int pinBl   = -1;
+    int8_t pinMosi = -1;
+    int8_t pinSclk = -1;
+    int8_t pinCs   = -1;
+    int8_t pinDc   = -1;
+    int8_t pinRst  = -1;
+    int8_t pinBl   = -1;
 
     bool initDone = false;
     bool lastPowerState = true;
 
-    // VERIFIED WORKING KEY DECLARATIONS WITH BRACKETS
     const char SETTING_PROFILE[] = "Hardware-Profile";
     const char SETTING_WIDTH[]   = "Display-Width";
     const char SETTING_HEIGHT[]  = "Display-Height";
@@ -97,7 +96,7 @@ class TTGO_TDISPLAY_OUTPUT : public Usermod {
       if (initDone) return;
 
       if (pinMosi >= 0) {
-        int8_t pinsToAllocate[] = { (int8_t)pinMosi, (int8_t)pinSclk, (int8_t)pinCs, (int8_t)pinDc, (int8_t)pinRst, (int8_t)pinBl };
+        int8_t pinsToAllocate[] = { pinMosi, pinSclk, pinCs, pinDc, pinRst, pinBl };
         for (uint8_t i = 0; i < 6; i++) {
           if (pinsToAllocate[i] >= 0) {
             PinManager::allocatePin(pinsToAllocate[i], false, PinOwner::UM_Unspecified);
